Removed return of terminating token when overshooting expr length
When overshooting expression length, the expressionbuilder returns
a terminating token, this is so if a function expects a token
there will be a token to evaluate. However, this may lead to
infinite loops so just one terminating token was added and if the
expression builder goes over this then the program will exit.